file writer done

+++ serial.h
@@ -7,6 +7,6 @@
 #include "linked_list.h"
 #define MAX 65535
 
-void serial_run(float m_member, float m_insert, float  m_delete, struct list_node** head, int m );
+void serial_run(int m_member, int m_insert, int m_delete, struct list_node **head, int m);
 
 #endif //LAB_SERIAL_H
